# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DailySpeak API",
    "version": "1.3.0",
    "description": "HTTP API for DailySpeak. Mobile and web clients share `/api/v1/auth/*`, short-lived Bearer access tokens, and rotating opaque refresh tokens. Native clients receive refresh tokens in JSON; browser requests use a scoped HttpOnly refresh cookie. Remaining unversioned web resource routes also require Bearer access tokens while they migrate to `/api/v1`."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current API origin"
    },
    {
      "url": "http://localhost:3219",
      "description": "Local API HTTP"
    },
    {
      "url": "https://localhost:3444",
      "description": "Local API HTTPS"
    },
    {
      "url": "https://api.example.com",
      "description": "Production API example"
    }
  ],
...
  "paths": {
...
    "/api/v1/recordings": {
      "get": {
        "tags": [
          "Mobile API v1",
          "Recordings"
        ],
        "summary": "List owned recordings with cursor pagination",
        "description": "Returns recordings ordered by timestamp and id descending. Cursors are opaque and must be sent back unchanged.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/PageLimitQuery"
          },
          {
            "$ref": "#/components/parameters/PageCursorQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/V1RecordingPageOK"
          },
          "400": {
            "$ref": "#/components/responses/V1BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/V1Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/V1MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/V1InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Mobile API v1",
          "Recordings"
        ],
        "summary": "Create a recording from ready media assets",
        "description": "Atomically attaches owned ready audio and optional photo assets, reserves quota, and enqueues durable processing. A user access token and an Idempotency-Key are required; guest identities cannot start paid processing.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordingFromAssetsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/V1RecordingWithQuotaCreated"
          },
          "400": {
            "$ref": "#/components/responses/V1BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/V1Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/V1Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/V1NotFound"
          },
          "409": {
            "$ref": "#/components/responses/V1Conflict"
          },
          "500": {
            "$ref": "#/components/responses/V1InternalServerError"
          }
        }
      }
    },
...
    "/api/v1/media/uploads": {
      "post": {
        "tags": [
          "Mobile API v1",
          "Media"
        ],
        "summary": "Create an owned resumable media upload",
        "description": "Creates an upload plan for recording audio or a recording photo. The client computes the complete-object SHA-256 before starting. User identities use the normal media limits. A guest identity may create one recording_audio upload up to 10 MiB for its bounded preview; guest photos and additional live assets are rejected.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestIdHeader"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/V1MediaUploadCreated"
          },
          "400": {
            "$ref": "#/components/responses/V1BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/V1Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/V1Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/V1Conflict"
          },
          "413": {
            "$ref": "#/components/responses/V1PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/V1UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/V1UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/V1InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/V1ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/daily-questions": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Generate daily speaking questions",
        "description": "Uses an optional Bearer access token. If authenticated, the user's saved English level overrides the `level` query parameter.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DateQuery"
          },
          {
            "$ref": "#/components/parameters/RefreshQuery"
          },
          {
            "$ref": "#/components/parameters/InterestQuery"
          },
          {
            "$ref": "#/components/parameters/EnglishLevelQuery"
          },
          {
            "name": "avoid",
            "in": "query",
            "description": "Questions to avoid; may be repeated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "Generated questions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyQuestionsResponse"
                },
                "example": {
                  "questions": [
                    "How do you usually plan a short trip?",
                    "What technology helps you study better?",
                    "How do you keep a healthy morning routine?"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/topic-guidance": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Generate follow-up questions and words for a topic",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 300
            }
          },
          {
            "$ref": "#/components/parameters/RefreshQuery"
          },
          {
            "$ref": "#/components/parameters/InterestQuery"
          },
          {
            "$ref": "#/components/parameters/EnglishLevelQuery"
          },
          {
            "name": "avoidQuestion",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "avoidWord",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "Topic guidance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicGuidanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/study-words": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "Generate a vocabulary pack",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RefreshQuery"
          },
          {
            "$ref": "#/components/parameters/InterestQuery"
          },
          {
            "$ref": "#/components/parameters/EnglishLevelQuery"
          },
          {
            "name": "avoidWord",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "Study words and context text.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyWordsResponse"
                }
              }
            }
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/user/data": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Load user profile data, recordings, quota and subscription",
        "responses": {
          "200": {
            "description": "User data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/user/interests": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Replace selected user interests",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInterestsRequest"
              },
              "example": {
                "interestIds": [
                  "technology",
                  "travel",
                  "fitness"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved interests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveInterestsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/user/ollama-model": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Read configured Ollama model",
        "responses": {
          "200": {
            "description": "Configured model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OllamaModelResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/user/subscription": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Read subscription and recording quota",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SubscriptionOK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Activate or extend subscription by one month",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SubscriptionOK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Mark active subscription as cancelled",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SubscriptionOK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/user/english-level": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Read saved English level",
        "responses": {
          "200": {
            "description": "Current level.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnglishLevelResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Save English level",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnglishLevelResponse"
              },
              "example": {
                "level": "b2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved level.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnglishLevelResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/user/recordings": {
      "post": {
        "tags": [
          "Recordings"
        ],
        "summary": "Save a full base64 recording for background processing",
        "description": "Legacy/fallback upload flow. Backend saves the recording immediately, returns it with `status=processing`, then runs transcription, AI suggestions, and the natural conversational rewrite in separate background steps.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/RecordingWithQuotaCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token returned by `/api/v1/auth/*`. Native refresh tokens are sent only to the refresh endpoint; browsers use a scoped HttpOnly refresh cookie."
      }
    },
    "parameters": {
      "RequestIdHeader": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Caller-generated diagnostic id. Unsafe values are replaced by the server.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$"
        }
      },
...
      "PageLimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Number of resources to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}